Renamed data provider files to match new convention, other lint fixes
