+++ src/renderer/hooks/papi-hooks/index.ts
@@ -1,8 +1,8 @@
 import usePromise from '@renderer/hooks/papi-hooks/use-promise.hook';
 import useEvent from '@renderer/hooks/papi-hooks/use-event.hook';
-import useEventAsync from '@renderer/hooks/papi-hooks/useEventAsync';
-import useDataProvider from '@renderer/hooks/papi-hooks/useDataProvider';
-import useData from '@renderer/hooks/papi-hooks/useData';
+import useEventAsync from '@renderer/hooks/papi-hooks/use-event-async.hook';
+import useDataProvider from '@renderer/hooks/papi-hooks/use-data-provider.hook';
+import useData from '@renderer/hooks/papi-hooks/use-data.hook';
 
 /** All React hooks to be exposed on the papi */
 const papiHooks = {
