+++ src/renderer/testing/test-buttons-panel.component.tsx
@@ -8,9 +8,9 @@ import { TabInfo } from '@shared/data/web-view.model';
 import { WebView, WebViewProps } from '@renderer/components/web-view.component';
 import useEvent from '@renderer/hooks/papi-hooks/use-event.hook';
 import { AddWebViewEvent } from '@shared/services/web-view.service';
-import useData from '@renderer/hooks/papi-hooks/useData';
-import useDataProvider from '@renderer/hooks/papi-hooks/useDataProvider';
-import IDataProvider from '@shared/models/IDataProvider';
+import useData from '@renderer/hooks/papi-hooks/use-data.hook';
+import useDataProvider from '@renderer/hooks/papi-hooks/use-data-provider.hook';
+import IDataProvider from '@shared/models/data-provider.interface';
 
 const testBase: (message: string) => Promise<string> =
   networkService.createRequestFunction('electronAPI.env.test');
