+++ src/shared/models/data-provider.interface.ts
